+++ src/Polyfill.NUnit/IsExtensions.cs
@@ -1,9 +1,10 @@
 ﻿using NUnit.Framework.Constraints;
+using System.Collections.Generic;
 
 namespace NUnit.Framework
 {
     /// <summary>
-    /// Provides extension methods for the <see cref="Is"/> class to polyfill constraints.
+    ///  Provides extension methods for the <see cref="Is"/> class to polyfill constraints.
     /// </summary>
     public static class IsExtensions
     {
@@ -18,6 +19,17 @@ namespace NUnit.Framework
             public static EqualConstraint EqualTo<T>(T[] expected)
             {
                 return new(expected);
+            }
+
+            /// <summary>
+            ///  Returns a constraint that tests whether the actual collection is equivalent to the expected collection.
+            /// </summary>
+            /// <typeparam name="T">The type of the collection elements.</typeparam>
+            /// <param name="expected">The expected collection.</param>
+            /// <returns>An <see cref="CollectionEquivalentConstraint"/> for the expected collection.</returns>
+            public static CollectionEquivalentConstraint EquivalentTo<T>(IEnumerable<T> expected)
+            {
+                return new CollectionEquivalentConstraint(expected);
             }
         }
     }