+++ src/Tests/Polyfill.NUnit.UnitTests/IsExtensionsTest.cs
@@ -12,5 +12,15 @@
                 Is.EqualTo([1, 2, 3])
             );
         }
+
+        [Test]
+        public void EquivalentTo_GenericOverloadForEnumerables()
+        {
+            var data = new[] { 1, 2, 3 };
+            Assert.That(
+                data,
+                Is.EquivalentTo([3, 2, 1])
+            );
+        }
     }
 }