+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ﻿# Polyfill.NUnit
 
-## 01.00.00 (2025.11.23)
+## 01.00.00 (2025.12.13)
 
 * Add C# 14 extension method for `Is.EqualTo<T>(T[])`
+* Add C# 14 extension method for `Is.EquivalentTo<T>(IEnumerable<T>])`
+* Add C# 14 extension method for `Assert.EnterMultipleScope()`